requirements.txt
python-chess
pygame
